component: create useFetch component
