+++ README.md
@@ -38,3 +38,4 @@ To add your own TMDB API Key, follow these steps:
 - [How to use Axios with React](https://www.freecodecamp.org/news/how-to-use-axios-with-react/)
 - [Get Requests in React using Async/Await](https://dev.to/redeyemedia/get-requests-in-react-using-async-await-7jd)
 - [Text Gradient with Font Awesome](https://stackoverflow.com/questions/12732663/text-gradient-with-font-awesome)
+- [useFetch Component](https://www.youtube.com/watch?v=j942wKiXFu8&list=PL4cUxeGkcC9gZD-Tvwfod2gaISzfRiP9d&index=1)
